+++ Data/Primitive.hs
@@ -15,6 +15,8 @@ module Data.Primitive (
   module Data.Primitive.Array,
   module Data.Primitive.ByteArray,
   module Data.Primitive.Addr,
+  module Data.Primitive.SmallArray,
+  module Data.Primitive.UnliftedArray,
 
   sizeOf, alignment
 ) where
@@ -23,3 +25,5 @@ import Data.Primitive.Types
 import Data.Primitive.Array
 import Data.Primitive.ByteArray
 import Data.Primitive.Addr
+import Data.Primitive.SmallArray
+import Data.Primitive.UnliftedArray